Add unit test support for AngularJS.
Installs Karma with PhantomJS (headless browser) plugin,
coverage support, and the Jasmine testing framework.
Also installs angular-mocks to support mocking out
Angular dependencies that would otherwise do silly things
like hit a server or try to be asynchronous, when we really
want tests to be synchronous if we can help it.